fix warning in ucol_res.cpp/isAcceptableUCA()
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/8171


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@30000 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/ucol_imp.h
@@ -182,7 +182,7 @@
 #define UCA_DATA_FORMAT_3 ((uint8_t)0x6c)
 
 #define UCA_FORMAT_VERSION_0 ((uint8_t)3)
-#define UCA_FORMAT_VERSION_1 ((uint8_t)0)
+#define UCA_FORMAT_VERSION_1 0
 #define UCA_FORMAT_VERSION_2 ((uint8_t)0)
 #define UCA_FORMAT_VERSION_3 ((uint8_t)0)
